// solid-csr-spa-template/src/generated/types/delete-profile-picture-response.ts
// Generated from rust-be-template OpenAPI. Do not edit by hand.


export type DeleteProfilePictureResponse = {
  readonly active_profile_picture_id: string | null;
  readonly cleanup_deleted_count: number;
  readonly cleanup_failure_count: number;
  readonly cleanup_remaining_count: number;
  readonly deleted_profile_picture_id: string;
};