// solid-csr-spa-template/src/generated/types/profile-picture-history-item.ts
// Generated from rust-be-template OpenAPI. Do not edit by hand.


export type ProfilePictureHistoryItem = {
  readonly created_at: string;
  readonly image_type: number;
  readonly is_active: boolean;
  readonly object_url: string;
  readonly profile_picture_id: string;
  readonly updated_at: string;
};